blog project/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-filter==25.2
markdown==3.9
Pillow==12.0.0
whitenoise==6.11.0
psycopg2-binary==2.9.11
